[refactor] 모임 메이트 추가 페이지를 모달로 변경
