+++ src/components/gathering/AddNewMateInGatheringModal.tsx
@@ -1,36 +1,31 @@
-import {useNavigation} from '@react-navigation/native';
 import React, {useEffect, useState} from 'react';
 import styled from 'styled-components/native';
-import {RootStackNavigationProp} from '~/App';
 import {
   responseFont as rf,
+  heightSizePercentage as hp,
   widthSizePercentage as wp,
 } from '~/components/common/ResponsiveSize';
-import BackView from '~/components/common/BackView';
 import {Keyboard} from 'react-native';
 import {showToast} from '~/components/common/modal/toastConfig';
 import LoadingModal from '~/components/common/modal/LoadingModal';
 import SearchExhFrame from '~/components/exhSearch/SearchExhFrame';
-import {checkBlankInKeyword} from '~/utils/keyword';
 import {useAddNewMateInGathering} from '~/api/queries/gathering';
 import {useEnterGatheringInfo} from '~/zustand/gathering/enterGathering';
-import SearchNewMateListInGathering from './SearchNewMateListInGathering';
+import {DEFAULT_TEXT, LIGHT_GREY, MAIN_COLOR} from '~/components/common/colors';
 import {
-  BACK_COLOR,
-  DEFAULT_TEXT,
-  LIGHT_GREY,
-  MAIN_COLOR,
-} from '~/components/common/colors';
-import {
-  AREA_FONT_SIZE,
   BUTTON_FONT_SIZE,
   BUTTON_PADDING,
   BUTTON_RADIUS,
   FONT_NAME,
 } from '~/components/common/style';
+import SearchNewMateListInGathering from './SearchNewMateListInGathering';
+import InfoModal from '../common/modal/InfoModal';
 
-const AddNewMateInGatheringScreen = () => {
-  const navigation = useNavigation<RootStackNavigationProp>();
+interface Props {
+  handleCloseModal: () => void;
+}
+
+const AddNewMateInGatheringModal: React.FC<Props> = ({handleCloseModal}) => {
   const {enterGatheringInfo} = useEnterGatheringInfo();
   const [isLoadingOpen, setIsLoadingOpen] = useState<boolean>(false);
   const [nicknameKeyword, setNicknameKeyword] = useState<string>('');
@@ -55,13 +50,12 @@ const AddNewMateInGatheringScreen = () => {
     }
     if (isSuccess) {
       showToast('전시 메이트 추가 성공 :)');
-      navigation.goBack();
+      handleCloseModal();
     }
   }, [isError, isLoading, isSuccess]);
 
   const onPressCreate = () => {
     Keyboard.dismiss();
-    console.log(selectedMate);
     if (selectedMate !== -1) {
       addNewMate();
     } else {
@@ -69,40 +63,37 @@ const AddNewMateInGatheringScreen = () => {
     }
   };
 
-  const onPressSearch = () => {
-    if (checkBlankInKeyword(nicknameKeyword)) {
-      showToast('다시 검색해 주세요.');
-    } else {
-      setSelectedMate(-1);
-      setKeyword(nicknameKeyword);
+  useEffect(() => {
+    if (nicknameKeyword === '') {
+      setKeyword('');
     }
+  }, [nicknameKeyword]);
+
+  const onPressSearch = () => {
+    setSelectedMate(-1);
+    setKeyword(nicknameKeyword);
     Keyboard.dismiss();
   };
 
   return (
-    <Container>
-      {/* header */}
-      <BackView title="모임 메이트 추가" line={true} />
-      {/* body */}
+    <InfoModal handleCloseModal={handleCloseModal}>
+      <AreaView>
+        <Message>모임 메이트 선택</Message>
+        <Message greyColor={true}>(내 전시 메이트만 가능)</Message>
+      </AreaView>
       <Contents>
-        <AreaView>
-          <AreaText>모임 메이트 선택</AreaText>
-          <AreaText greyColor={true}>(내 전시 메이트만 가능)</AreaText>
-        </AreaView>
         <SearchExhFrame
           searchKeyword={nicknameKeyword}
           onPressSearch={onPressSearch}
           handleSearchKeyword={setNicknameKeyword}
           searchMessage={'닉네임을 검색하세요'}>
           {/* 모임 메이트 목록 */}
-          {keyword !== '' && (
-            <SearchNewMateListInGathering
-              searchKeyword={keyword}
-              changeIsPressed={onPressSearch}
-              selectedMate={selectedMate}
-              handleSelectedMate={setSelectedMate}
-            />
-          )}
+          <SearchNewMateListInGathering
+            searchKeyword={keyword}
+            changeIsPressed={onPressSearch}
+            selectedMate={selectedMate}
+            handleSelectedMate={setSelectedMate}
+          />
         </SearchExhFrame>
         {/* 모임 메이트 추가 버튼 */}
         <ButtonTouch
@@ -115,23 +106,17 @@ const AddNewMateInGatheringScreen = () => {
         </ButtonTouch>
       </Contents>
       {isLoadingOpen && <LoadingModal message={'모임 메이트 추가 중'} />}
-    </Container>
+    </InfoModal>
   );
 };
 
-export default AddNewMateInGatheringScreen;
+export default AddNewMateInGatheringModal;
 
 /** style */
-const Container = styled.View`
-  flex: 1;
-  flex-direction: column;
-  background-color: ${BACK_COLOR};
-`;
-
 const AreaView = styled.View`
   flex-direction: row;
-  padding-top: ${wp(3.3)}px;
-  padding-left: ${wp(4.8)}px;
+  padding-top: ${hp(1)}px;
+  padding-left: ${wp(5)}px;
   gap: ${wp(1.5)}px;
   align-items: center;
 `;
@@ -140,9 +125,9 @@ interface AreaTextProps {
   greyColor: boolean;
 }
 
-const AreaText = styled.Text<AreaTextProps>`
+const Message = styled.Text`
   font-size: ${(props: AreaTextProps) =>
-    props.greyColor ? `${rf(14)}px` : `${AREA_FONT_SIZE}px`};
+    props.greyColor ? `${rf(14)}px` : `${BUTTON_FONT_SIZE}px`};
   color: ${(props: AreaTextProps) =>
     props.greyColor ? `${LIGHT_GREY}` : `${DEFAULT_TEXT}`};
   font-family: ${FONT_NAME};
@@ -150,7 +135,6 @@ const AreaText = styled.Text<AreaTextProps>`
 
 const Contents = styled.View`
   flex: 1;
-  padding-bottom: ${wp(3.3)}px;
 `;
 
 const ButtonTouch = styled.TouchableOpacity`
